+++ mobile/app/src/main/java/com/lacuc/pets/data/animal/AnimalRepository.kt
@@ -1,5 +1,9 @@
 package com.lacuc.pets.data.animal
 
+import com.lacuc.pets.data.animal.entity.Animal
+import com.lacuc.pets.data.animal.entity.Medical
+import com.lacuc.pets.data.animal.entity.Memo
+
 interface AnimalRepository {
     fun loadAnimal(gid: Int): List<Animal>
