--- mobile/app/src/main/java/com/lacuc/pets/data/animal/Memo.kt
@@ -1,3 +1,0 @@
-package com.lacuc.pets.data.animal
-
-data class Memo(val content: String)
