+++ mobile/app/src/main/java/com/lacuc/pets/data/animal/entity/Medical.kt
@@ -1,4 +1,4 @@
-package com.lacuc.pets.data.animal
+package com.lacuc.pets.data.animal.entity
 
 data class Medical(
     val date: Long,
