+++ mobile/app/src/main/java/com/lacuc/pets/data/animal/entity/Memo.kt
@@ -1,0 +1,3 @@
+package com.lacuc.pets.data.animal.entity
+
+data class Memo(val content: String)
