+++ mobile/app/src/main/java/com/lacuc/pets/domain/animal/memo/AddMemoUseCase.kt
@@ -1,7 +1,7 @@
 package com.lacuc.pets.domain.animal.memo
 
 import com.lacuc.pets.data.animal.AnimalRepository
-import com.lacuc.pets.data.animal.Memo
+import com.lacuc.pets.data.animal.entity.Memo
 import javax.inject.Inject
 
 class AddMemoUseCase @Inject constructor(private val repository: AnimalRepository) {
