+++ mobile/app/src/main/java/com/lacuc/pets/ui/manage/animal/detail/memo/AddMemoViewModel.kt
@@ -2,7 +2,7 @@ package com.lacuc.pets.ui.manage.animal.detail.memo
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.lacuc.pets.data.animal.Memo
+import com.lacuc.pets.data.animal.entity.Memo
 import com.lacuc.pets.domain.animal.memo.AddMemoUseCase
 import com.lacuc.pets.util.SingleLiveEvent
 import com.lacuc.pets.util.safeValue
