+++ mobile/app/src/main/res/navigation/nav_manage.xml
@@ -66,7 +66,7 @@
         tools:layout="@layout/fragment_add_animal">
         <argument
             android:name="animal"
-            app:argType="com.lacuc.pets.data.animal.Animal"
+            app:argType="com.lacuc.pets.data.animal.entity.Animal"
             app:nullable="true" />
     </fragment>
 
@@ -77,7 +77,7 @@
         tools:layout="@layout/fragment_animal_detail">
         <argument
             android:name="animal"
-            app:argType="com.lacuc.pets.data.animal.Animal" />
+            app:argType="com.lacuc.pets.data.animal.entity.Animal" />
         <action
             android:id="@+id/action_animalDetailFragment_to_addAnimalFragment"
             app:destination="@id/addAnimalFragment" />
